Fix CS0121: cast CurrentLevel to int to resolve Math.Min overload ambiguity

+++ PKHeX.Mobile/Pages/BankViewPage.xaml.cs
@@ -882,7 +882,7 @@ public partial class BankViewPage : ContentPage
         try
         {
             var learnset = LearnSource9SV.Instance.GetLearnset((ushort)pk.Species, 0);
-            byte level = (byte)Math.Min(pk.CurrentLevel, 100);
+            byte level = (byte)Math.Min((int)pk.CurrentLevel, 100);
             var moves = learnset.GetMoveRange(level);
             if (moves.IsEmpty)
                 moves = learnset.GetAllMoves(); // fallback: full learnset
